pc配置中心-频道配置100%

+++ core/src/main/java/com/smart/home/modules/article/service/ArticleChannelService.java
@@ -1,9 +1,14 @@
 package com.smart.home.modules.article.service;
 
 import com.github.pagehelper.PageHelper;
+import com.smart.home.common.enums.RecordStatusEnum;
+import com.smart.home.common.exception.DuplicateDataException;
+import com.smart.home.common.exception.ServiceException;
 import com.smart.home.modules.article.dao.ArticleChannelMapper;
+import com.smart.home.modules.article.dao.ArticleMapper;
 import com.smart.home.modules.article.entity.ArticleChannel;
 import com.smart.home.modules.article.entity.ArticleChannelExample;
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -19,13 +24,29 @@ public class ArticleChannelService {
 
     @Resource
     ArticleChannelMapper articleChannelMapper;
+    @Resource
+    ArticleMapper articleMapper;
 
     public int create(ArticleChannel articleChannel) {
+        // 检查同名频道
+        ArticleChannelExample example = new ArticleChannelExample();
+        example.createCriteria().andTitleEqualTo(articleChannel.getTitle());
+        if (articleChannelMapper.countByExample(example) > 0) {
+            throw new DuplicateDataException("该频道已经存在");
+        }
         articleChannel.setCreatedTime(new Date());
+        articleChannel.setRevision(0);
+        articleChannel.setState(RecordStatusEnum.NORMAL.getStatus());
         return articleChannelMapper.insertSelective(articleChannel);
     }
 
     public int update(ArticleChannel articleChannel) {
+        // 查看变更后的名称有没有冲突
+        ArticleChannelExample example = new ArticleChannelExample();
+        example.createCriteria().andTitleEqualTo(articleChannel.getTitle()).andIdNotEqualTo(articleChannel.getId());
+        if (articleChannelMapper.countByExample(example) > 0) {
+            throw new DuplicateDataException("该频道已经存在");
+        }
         articleChannel.setUpdatedTime(new Date());
         return articleChannelMapper.updateByPrimaryKeySelective(articleChannel);
     }
@@ -37,6 +58,10 @@ public class ArticleChannelService {
     @Transactional(rollbackFor = RuntimeException.class)
     public void delete(List<Long> idList) {
         for (Long id : idList) {
+            // 检查有没有关联的文章了，有的话不能删除
+            if (articleMapper.countArticleByChannelId(id) > 0) {
+                throw new ServiceException("该频道有关联的文章了，不能删除");
+            }
             articleChannelMapper.deleteByPrimaryKey(id.intValue());
         }
     }
@@ -45,7 +70,9 @@ public class ArticleChannelService {
         PageHelper.startPage(pageNum, pageSize);
         ArticleChannelExample example = new ArticleChannelExample();
         ArticleChannelExample.Criteria criteria = example.createCriteria();
-        // TODO 按需根据字段查询
+        if (StringUtils.isNotBlank(articleChannel.getTitle())) {
+            criteria.andTitleEqualTo(articleChannel.getTitle());
+        }
         return articleChannelMapper.selectByExample(example);
     }
 
@@ -54,4 +81,10 @@ public class ArticleChannelService {
         return articleChannel;
     }
 
+    public List<ArticleChannel> selectAllValid() {
+        ArticleChannelExample example = new ArticleChannelExample();
+        example.createCriteria().andStateEqualTo(RecordStatusEnum.NORMAL.getStatus());
+        example.setOrderByClause("sort desc");
+        return articleChannelMapper.selectByExample(example);
+    }
 }
